# TSDB Index reuses slices, adds pools