--- a/pkg/storage/tsdb/multi_file_index_test.go
+++ b/pkg/storage/tsdb/multi_file_index_test.go
@@ -67,8 +67,7 @@ func TestMultiIndex(t *testing.T) {
 	require.Nil(t, err)
 
 	t.Run("GetChunkRefs", func(t *testing.T) {
-		refs := ChunkRefsPool.Get()
-		err := idx.GetChunkRefs(context.Background(), "fake", 2, 5, &refs, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
+		refs, err := idx.GetChunkRefs(context.Background(), "fake", 2, 5, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
 		require.Nil(t, err)
 
 		expected := []ChunkRef{
@@ -105,8 +104,7 @@ func TestMultiIndex(t *testing.T) {
 	})
 
 	t.Run("Series", func(t *testing.T) {
-		xs := SeriesPool.Get()
-		err := idx.Series(context.Background(), "fake", 2, 5, &xs, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
+		xs, err := idx.Series(context.Background(), "fake", 2, 5, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
 		require.Nil(t, err)
 		expected := []Series{
 			{