--- a/tools/tsdb/tsdb-map/main_test.go
+++ b/tools/tsdb/tsdb-map/main_test.go
@@ -76,8 +76,9 @@ func BenchmarkQuery_GetChunkRefs(b *testing.B) {
 		}
 		idx := tsdb.NewTSDBIndex(reader)
 		b.Run(bm.name, func(b *testing.B) {
+			refs := tsdb.ChunkRefsPool.Get()
 			for i := 0; i < b.N; i++ {
-				_, err := idx.GetChunkRefs(context.Background(), "fake", 0, math.MaxInt64, nil, bm.matchers...)
+				err := idx.GetChunkRefs(context.Background(), "fake", 0, math.MaxInt64, &refs, nil, bm.matchers...)
 				if err != nil {
 					panic(err)
 				}
@@ -101,6 +102,7 @@ func BenchmarkQuery_GetChunkRefsSharded(b *testing.B) {
 		shardFactor := 16
 
 		b.Run(bm.name, func(b *testing.B) {
+			refs := tsdb.ChunkRefsPool.Get()
 			for i := 0; i < b.N; i++ {
 				for j := 0; j < shardFactor; j++ {
 					shard := index.ShardAnnotation{
@@ -108,7 +110,7 @@ func BenchmarkQuery_GetChunkRefsSharded(b *testing.B) {
 						Of:    uint32(shardFactor),
 					}
 
-					_, err := idx.GetChunkRefs(context.Background(), "fake", 0, math.MaxInt64, &shard, bm.matchers...)
+					err := idx.GetChunkRefs(context.Background(), "fake", 0, math.MaxInt64, &refs, &shard, bm.matchers...)
 					if err != nil {
 						panic(err)
 					}