This PR changes the index interface to accept slices. This allows us to reuse memory more easily. To make this easier, I've introduced a number of releveant pools for us to use.

It also adds a configurable `rounds` parameter (default 8) for the benchmarking script.

ref https://github.com/grafana/loki/issues/5428

```
benchmarks:

name                                         old time/op    new time/op    delta
Query_PostingsForMatchers/match_ns-4           4.03µs ± 1%    4.02µs ± 0%     ~     (p=0.159 n=7+8)
Query_PostingsForMatchers/match_ns_regexp-4    4.03µs ± 0%    4.02µs ± 0%   -0.28%  (p=0.031 n=8+8)
Query_GetChunkRefs/match_ns-4                  27.4ms ± 0%    14.2ms ± 0%  -48.36%  (p=0.001 n=6+8)
Query_GetChunkRefs/match_ns_regexp-4           27.5ms ± 1%    14.3ms ± 1%  -47.97%  (p=0.000 n=8+7)
Query_GetChunkRefsSharded/match_ns-4            120ms ± 1%     102ms ± 1%  -15.29%  (p=0.000 n=8+8)
Query_GetChunkRefsSharded/match_ns_regexp-4     119ms ± 1%     102ms ± 1%  -14.64%  (p=0.000 n=7+8)

name                                         old alloc/op   new alloc/op   delta
Query_PostingsForMatchers/match_ns-4            80.0B ± 0%     80.0B ± 0%     ~     (all equal)
Query_PostingsForMatchers/match_ns_regexp-4     80.0B ± 0%     80.0B ± 0%     ~     (all equal)
Query_GetChunkRefs/match_ns-4                  40.4MB ± 0%     7.4MB ± 0%  -81.71%  (p=0.000 n=8+8)
Query_GetChunkRefs/match_ns_regexp-4           40.4MB ± 0%     7.4MB ± 0%  -81.70%  (p=0.000 n=8+7)
Query_GetChunkRefsSharded/match_ns-4           48.6MB ± 0%    15.0MB ± 0%  -69.20%  (p=0.001 n=7+7)
Query_GetChunkRefsSharded/match_ns_regexp-4    48.6MB ± 0%    15.0MB ± 0%  -69.20%  (p=0.001 n=8+6)

name                                         old allocs/op  new allocs/op  delta
Query_PostingsForMatchers/match_ns-4             4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Query_PostingsForMatchers/match_ns_regexp-4      4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Query_GetChunkRefs/match_ns-4                    278k ± 0%      278k ± 0%   -0.01%  (p=0.000 n=8+8)
Query_GetChunkRefs/match_ns_regexp-4             278k ± 0%      278k ± 0%   -0.01%  (p=0.000 n=8+8)
Query_GetChunkRefsSharded/match_ns-4             394k ± 0%      394k ± 0%   -0.12%  (p=0.000 n=8+8)
Query_GetChunkRefsSharded/match_ns_regexp-4      394k ± 0%      394k ± 0%   -0.12%  (p=0.000 n=8+8)
```